docker_streamlit/requirements.txt
streamlit==1.22.0
pandas==1.3.5
kafka-python==2.0.2